chore: Refactor macro - separate code by type of render
