+++ leptos_macro/src/view/mod.rs
@@ -1,0 +1,563 @@
+use crate::{attribute_value, Mode};
+use convert_case::{Case::Snake, Casing};
+
+use proc_macro2::{Ident, Span, TokenStream, TokenTree};
+use quote::{quote, quote_spanned};
+use rstml::node::{
+    KeyedAttribute, Node, NodeElement, NodeName,
+};
+
+use syn::{spanned::Spanned, Expr, Expr::Tuple, ExprLit, ExprPath, Lit};
+
+pub mod client_builder;
+pub mod client_template;
+pub mod server_template;
+pub mod component_builder;
+pub mod slot_helper;
+pub mod ide_helper;
+
+pub(crate) use ide_helper::*;
+
+
+
+pub(crate) fn render_view(
+    cx: &Ident,
+    nodes: &[Node],
+    mode: Mode,
+    global_class: Option<&TokenTree>,
+    call_site: Option<String>,
+) -> TokenStream {
+    let empty = {
+        let span = Span::call_site();
+        quote_spanned! {
+            span => leptos::leptos_dom::Unit
+        }
+    };
+
+    if mode == Mode::Ssr {
+        match nodes.len() {
+            0 => empty,
+            1 => {
+                server_template::root_node_to_tokens_ssr(cx, &nodes[0], global_class, call_site)
+            }
+            _ => server_template::fragment_to_tokens_ssr(
+                cx,
+                Span::call_site(),
+                nodes,
+                global_class,
+                call_site,
+            ),
+        }
+    } else {
+        match nodes.len() {
+            0 => empty,
+            1 => client_builder::node_to_tokens(
+                cx,
+                &nodes[0],
+                client_builder::TagType::Unknown,
+                None,
+                global_class,
+                call_site,
+            )
+            .unwrap_or_default(),
+            _ => client_builder::fragment_to_tokens(
+                cx,
+                Span::call_site(),
+                nodes,
+                true,
+                client_builder::TagType::Unknown,
+                None,
+                global_class,
+                call_site,
+            )
+            .unwrap_or(empty),
+        }
+    }
+}
+
+const TYPED_EVENTS: [&str; 126] = [
+    "afterprint",
+    "beforeprint",
+    "beforeunload",
+    "gamepadconnected",
+    "gamepaddisconnected",
+    "hashchange",
+    "languagechange",
+    "message",
+    "messageerror",
+    "offline",
+    "online",
+    "pagehide",
+    "pageshow",
+    "popstate",
+    "rejectionhandled",
+    "storage",
+    "unhandledrejection",
+    "unload",
+    "abort",
+    "animationcancel",
+    "animationend",
+    "animationiteration",
+    "animationstart",
+    "auxclick",
+    "beforeinput",
+    "blur",
+    "canplay",
+    "canplaythrough",
+    "change",
+    "click",
+    "close",
+    "compositionend",
+    "compositionstart",
+    "compositionupdate",
+    "contextmenu",
+    "cuechange",
+    "dblclick",
+    "drag",
+    "dragend",
+    "dragenter",
+    "dragleave",
+    "dragover",
+    "dragstart",
+    "drop",
+    "durationchange",
+    "emptied",
+    "ended",
+    "error",
+    "focus",
+    "focusin",
+    "focusout",
+    "formdata",
+    "gotpointercapture",
+    "input",
+    "invalid",
+    "keydown",
+    "keypress",
+    "keyup",
+    "load",
+    "loadeddata",
+    "loadedmetadata",
+    "loadstart",
+    "lostpointercapture",
+    "mousedown",
+    "mouseenter",
+    "mouseleave",
+    "mousemove",
+    "mouseout",
+    "mouseover",
+    "mouseup",
+    "pause",
+    "play",
+    "playing",
+    "pointercancel",
+    "pointerdown",
+    "pointerenter",
+    "pointerleave",
+    "pointermove",
+    "pointerout",
+    "pointerover",
+    "pointerup",
+    "progress",
+    "ratechange",
+    "reset",
+    "resize",
+    "scroll",
+    "securitypolicyviolation",
+    "seeked",
+    "seeking",
+    "select",
+    "selectionchange",
+    "selectstart",
+    "slotchange",
+    "stalled",
+    "submit",
+    "suspend",
+    "timeupdate",
+    "toggle",
+    "touchcancel",
+    "touchend",
+    "touchmove",
+    "touchstart",
+    "transitioncancel",
+    "transitionend",
+    "transitionrun",
+    "transitionstart",
+    "volumechange",
+    "waiting",
+    "webkitanimationend",
+    "webkitanimationiteration",
+    "webkitanimationstart",
+    "webkittransitionend",
+    "wheel",
+    "DOMContentLoaded",
+    "devicemotion",
+    "deviceorientation",
+    "orientationchange",
+    "copy",
+    "cut",
+    "paste",
+    "fullscreenchange",
+    "fullscreenerror",
+    "pointerlockchange",
+    "pointerlockerror",
+    "readystatechange",
+    "visibilitychange",
+];
+
+
+pub(crate) fn parse_event_name(name: &str) -> (TokenStream, bool, bool) {
+    let (name, is_force_undelegated) = parse_event(name);
+
+    let event_type = TYPED_EVENTS
+        .iter()
+        .find(|e| **e == name)
+        .copied()
+        .unwrap_or("Custom");
+    let is_custom = event_type == "Custom";
+
+    let Ok(event_type) = event_type.parse::<TokenStream>() else {
+            abort!(event_type, "couldn't parse event name");
+        };
+
+    let event_type = if is_custom {
+        quote! { Custom::new(#name) }
+    } else {
+        event_type
+    };
+    (event_type, is_custom, is_force_undelegated)
+}
+
+fn expr_to_ident(expr: &syn::Expr) -> Option<&ExprPath> {
+    match expr {
+        syn::Expr::Block(block) => block.block.stmts.last().and_then(|stmt| {
+            if let syn::Stmt::Expr(expr, ..) = stmt {
+                expr_to_ident(expr)
+            } else {
+                None
+            }
+        }),
+        syn::Expr::Path(path) => Some(path),
+        _ => None,
+    }
+}
+
+fn convert_to_snake_case(name: String) -> String {
+    if !name.is_case(Snake) {
+        name.to_case(Snake)
+    } else {
+        name
+    }
+}
+
+fn is_custom_element(tag: &str) -> bool {
+    tag.contains('-')
+}
+
+fn is_self_closing(node: &NodeElement) -> bool {
+    // self-closing tags
+    // https://developer.mozilla.org/en-US/docs/Glossary/Empty_element
+    matches!(
+        node.name().to_string().as_str(),
+        "area"
+            | "base"
+            | "br"
+            | "col"
+            | "embed"
+            | "hr"
+            | "img"
+            | "input"
+            | "link"
+            | "meta"
+            | "param"
+            | "source"
+            | "track"
+            | "wbr"
+    )
+}
+
+fn camel_case_tag_name(tag_name: &str) -> String {
+    let mut chars = tag_name.chars();
+    let first = chars.next();
+    let underscore = if tag_name == "option" || tag_name == "use" {
+        "_"
+    } else {
+        ""
+    };
+    first
+        .map(|f| f.to_ascii_uppercase())
+        .into_iter()
+        .chain(chars)
+        .collect::<String>()
+        + underscore
+}
+
+fn is_svg_element(tag: &str) -> bool {
+    matches!(
+        tag,
+        "animate"
+            | "animateMotion"
+            | "animateTransform"
+            | "circle"
+            | "clipPath"
+            | "defs"
+            | "desc"
+            | "discard"
+            | "ellipse"
+            | "feBlend"
+            | "feColorMatrix"
+            | "feComponentTransfer"
+            | "feComposite"
+            | "feConvolveMatrix"
+            | "feDiffuseLighting"
+            | "feDisplacementMap"
+            | "feDistantLight"
+            | "feDropShadow"
+            | "feFlood"
+            | "feFuncA"
+            | "feFuncB"
+            | "feFuncG"
+            | "feFuncR"
+            | "feGaussianBlur"
+            | "feImage"
+            | "feMerge"
+            | "feMergeNode"
+            | "feMorphology"
+            | "feOffset"
+            | "fePointLight"
+            | "feSpecularLighting"
+            | "feSpotLight"
+            | "feTile"
+            | "feTurbulence"
+            | "filter"
+            | "foreignObject"
+            | "g"
+            | "hatch"
+            | "hatchpath"
+            | "image"
+            | "line"
+            | "linearGradient"
+            | "marker"
+            | "mask"
+            | "metadata"
+            | "mpath"
+            | "path"
+            | "pattern"
+            | "polygon"
+            | "polyline"
+            | "radialGradient"
+            | "rect"
+            | "set"
+            | "stop"
+            | "svg"
+            | "switch"
+            | "symbol"
+            | "text"
+            | "textPath"
+            | "tspan"
+            | "use"
+            | "use_"
+            | "view"
+    )
+}
+
+fn is_math_ml_element(tag: &str) -> bool {
+    matches!(
+        tag,
+        "math"
+            | "mi"
+            | "mn"
+            | "mo"
+            | "ms"
+            | "mspace"
+            | "mtext"
+            | "menclose"
+            | "merror"
+            | "mfenced"
+            | "mfrac"
+            | "mpadded"
+            | "mphantom"
+            | "mroot"
+            | "mrow"
+            | "msqrt"
+            | "mstyle"
+            | "mmultiscripts"
+            | "mover"
+            | "mprescripts"
+            | "msub"
+            | "msubsup"
+            | "msup"
+            | "munder"
+            | "munderover"
+            | "mtable"
+            | "mtd"
+            | "mtr"
+            | "maction"
+            | "annotation"
+            | "semantics"
+    )
+}
+
+fn is_ambiguous_element(tag: &str) -> bool {
+    tag == "a" || tag == "script" || tag == "title"
+}
+
+fn parse_event(event_name: &str) -> (&str, bool) {
+    if let Some(event_name) = event_name.strip_suffix(":undelegated") {
+        (event_name, true)
+    } else {
+        (event_name, false)
+    }
+}
+
+
+
+fn fancy_class_name<'a>(
+    name: &str,
+    cx: &Ident,
+    node: &'a KeyedAttribute,
+) -> Option<(TokenStream, String, &'a Expr)> {
+    // special case for complex class names:
+    // e.g., Tailwind `class=("mt-[calc(100vh_-_3rem)]", true)`
+    if name == "class" {
+        if let Some(Tuple(tuple)) = node.value() {
+            if tuple.elems.len() == 2 {
+                let span = node.key.span();
+                let class = quote_spanned! {
+                    span => .class
+                };
+                let class_name = &tuple.elems[0];
+                let class_name = if let Expr::Lit(ExprLit {
+                    lit: Lit::Str(s),
+                    ..
+                }) = class_name
+                {
+                    s.value()
+                } else {
+                    proc_macro_error::emit_error!(
+                        class_name.span(),
+                        "class name must be a string literal"
+                    );
+                    Default::default()
+                };
+                let value = &tuple.elems[1];
+                return Some((
+                    quote! {
+                        #class(#class_name, (#cx, #value))
+                    },
+                    class_name,
+                    value,
+                ));
+            } else {
+                proc_macro_error::emit_error!(
+                    tuple.span(),
+                    "class tuples must have two elements."
+                )
+            }
+        }
+    }
+    None
+}
+
+fn ident_from_tag_name(tag_name: &NodeName) -> Ident {
+    match tag_name {
+        NodeName::Path(path) => path
+            .path
+            .segments
+            .iter()
+            .last()
+            .map(|segment| segment.ident.clone())
+            .expect("element needs to have a name"),
+        NodeName::Block(_) => {
+            let span = tag_name.span();
+            proc_macro_error::emit_error!(
+                span,
+                "blocks not allowed in tag-name position"
+            );
+            Ident::new("", span)
+        }
+        _ => Ident::new(
+            &tag_name.to_string().replace(['-', ':'], "_"),
+            tag_name.span(),
+        ),
+    }
+}
+
+fn fancy_style_name<'a>(
+    name: &str,
+    cx: &Ident,
+    node: &'a KeyedAttribute,
+) -> Option<(TokenStream, String, &'a Expr)> {
+    // special case for complex dynamic style names:
+    if name == "style" {
+        if let Some(Tuple(tuple)) = node.value() {
+            if tuple.elems.len() == 2 {
+                let span = node.key.span();
+                let style = quote_spanned! {
+                    span => .style
+                };
+                let style_name = &tuple.elems[0];
+                let style_name = if let Expr::Lit(ExprLit {
+                    lit: Lit::Str(s),
+                    ..
+                }) = style_name
+                {
+                    s.value()
+                } else {
+                    proc_macro_error::emit_error!(
+                        style_name.span(),
+                        "style name must be a string literal"
+                    );
+                    Default::default()
+                };
+                let value = &tuple.elems[1];
+                return Some((
+                    quote! {
+                        #style(#style_name, (#cx, #value))
+                    },
+                    style_name,
+                    value,
+                ));
+            } else {
+                proc_macro_error::emit_error!(
+                    tuple.span(),
+                    "style tuples must have two elements."
+                )
+            }
+        }
+    }
+    None
+}
+
+pub(crate) fn event_from_attribute_node(
+    attr: &KeyedAttribute,
+    force_undelegated: bool,
+) -> (TokenStream, &Expr) {
+    let event_name = attr
+        .key
+        .to_string()
+        .strip_prefix("on:")
+        .expect("expected `on:` directive")
+        .to_owned();
+
+    let handler = attribute_value(attr);
+
+    #[allow(unused_variables)]
+    let (name, name_undelegated) = parse_event(&event_name);
+
+    let event_type = TYPED_EVENTS
+        .iter()
+        .find(|e| **e == name)
+        .copied()
+        .unwrap_or("Custom");
+
+    let Ok(event_type) = event_type.parse::<TokenStream>() else {
+        abort!(attr.key, "couldn't parse event name");
+    };
+
+    let event_type = if force_undelegated || name_undelegated {
+        quote! { ::leptos::leptos_dom::ev::undelegated(::leptos::leptos_dom::ev::#event_type) }
+    } else {
+        quote! { ::leptos::leptos_dom::ev::#event_type }
+    };
+    (event_type, handler)
+}
